rename: API 통신, 외부 서비스 연동하는 폴더명 변경 #22
