+++ app/services/openai.ts
